word-ranking/web/requirements.txt
flask
requests
bs4
mysql-connector-python
